% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_moment_calc.R
\name{tidy_calc_moment}
\alias{tidy_calc_moment}
\title{Calculate community weighted mean, variance, skewness, and kurtosis}
\usage{
tidy_calc_moment(df, trait_names, comm_names, trait_value, weight)
}
\arguments{
\item{df}{A data frame containing the trait and community data.}

\item{trait_names}{Unquoted column name containing the trait IDs.}

\item{comm_names}{Unquoted column name containing the community names/IDs.}

\item{trait_value}{Unquoted column name containing the actual trait values.}

\item{weight}{Unquoted column name containing the variable used to weight the trait values (e.g. abundance).}
}
\value{
Returns a tibble with the columns comm, trait, cwm (mean), cwv (variance), cws (skewness), and cwk (excess kurtosis).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Calculates the community weighted mean, variance, skewness, and excess kurtosis of a given trait based on the moments described in Wieczynski et al. (2019).
}
\examples{
df <- data.frame(trait = c("height", "height", "weight", "weight"),
                 trait_value = c(5, 10, 15, 12),
                 abundancia = c(1, 2, 1, 3),
                 comm = c("A", "A", "B", "B"))
tidy_calc_moment(df,
trait_names = trait,
comm_names = comm,
trait_value = trait_value,
weight = abundancia)
}
\references{
Enquist, B. J., Norberg, J., Bonser, S. P., Violle, C., Webb, C. T., Henderson, A., ... & Savage, V. M. (2015).
Scaling from traits to ecosystems. \emph{Advances in Ecological Research}, 52, 249-318.

Metcalfe, R. J., Ozturk, M., & Pouteau, R. (2020).
Using functional traits to model annual plant community dynamics. \emph{Ecology}.

\enc{Šímová}{Simova}, I., Violle, C., Kraft, N. J., Storch, D., Svenning, J. C., Gallagher, R. V., ... & Enquist, B. J. (2018).
Spatial patterns and climate relationships of major plant traits in the New World differ between woody and herbaceous species. \emph{Global Ecology and Biogeography}, 27(8), 895-916.

Wieczynski, D. J., Boyle, B., Buzzard, V., Duran, S. M., Henderson, A. N., Hulshof, C. M., ... & Savage, V. M. (2019).
Climate shapes and shifts functional biodiversity in forests worldwide. \emph{Proceedings of the National Academy of Sciences}, 116(2), 587-592.
}
