% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_calc_rarity.R
\name{tidy_calc_rarity}
\alias{tidy_calc_rarity}
\title{Calculate functional rarity indices from tidy data}
\usage{
tidy_calc_rarity(df, comm_names, sp_names, trait_names, trait_value, weight)
}
\arguments{
\item{df}{A data frame containing the trait and community data.}

\item{comm_names}{Unquoted column name containing the community names/IDs.}

\item{sp_names}{Unquoted column name containing the species names/IDs.}

\item{trait_names}{Unquoted column name containing the trait IDs.}

\item{trait_value}{Unquoted column name containing the actual trait values.}

\item{weight}{Unquoted column name containing the variable used to weight the trait values (e.g. abundance).}
}
\value{
A list of tibbles containing distinctiveness and uniqueness metrics.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Wraps the \code{funrar} package to calculate functional rarity indices
(Distinctiveness and Uniqueness) directly from a tidy long-format data frame.
}
