% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyttmoment-package.R, R/tidyttmoment.R
\docType{package}
\name{tidyttmoment-package}
\alias{tidyttmoment-package}
\title{tidyttmoment: Functional Trait Moment Calculation}
\description{
Calculates the community four 'moments' (mean, variance, skewness, and kurtosis) of a given trait based on the moments described in Wieczynski et al. (2019) \doi{10.1073/pnas.1813723116}. These functional metrics are extremely useful in characterizing the distribution of traits in a plant community. It also provides tidyverse-friendly wrappers to seamlessly calculate advanced functional diversity indices (e.g., FDis, Rao's Q) using 'fundiversity' (Grenie et al. 2023 \doi{10.1111/ecog.06585}) and functional rarity indices using 'funrar' (Grenie et al. 2017 \doi{10.1111/ddi.12629}). Evaluating these community-weighted moments and diversity metrics allows researchers to evaluate shifts in optimal phenotypes and understand ecological filtering with exactness.

Calculates the community four 'moments' (mean, variance, skewness, and kurtosis) of a given trait based on the moments described in Wieczynski et al. (2019) \doi{10.1073/pnas.1813723116}. These functional metrics are extremely useful in characterizing the distribution of traits in a plant community. It also provides tidyverse-friendly wrappers to seamlessly calculate advanced functional diversity indices (e.g., FDis, Rao's Q) using 'fundiversity' (Grenie et al. 2023 \doi{10.1111/ecog.06585}) and functional rarity indices using 'funrar' (Grenie et al. 2017 \doi{10.1111/ddi.12629}). Evaluating these community-weighted moments and diversity metrics allows researchers to evaluate shifts in optimal phenotypes and understand ecological filtering with exactness.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/tidyttmoment}
  \item \url{https://paulesantos.github.io/tidyttmoment/}
  \item Report bugs at \url{https://github.com/PaulESantos/tidyttmoment/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/tidyttmoment}
  \item \url{https://paulesantos.github.io/tidyttmoment/}
  \item Report bugs at \url{https://github.com/PaulESantos/tidyttmoment/issues}
}

}
\author{
\strong{Maintainer}: Paul Efren Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

Other contributors:
\itemize{
  \item Macroecology Lab [copyright holder, funder]
}

}
\keyword{internal}
