% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttmoment-package.R, R/ttmoment.R
\docType{package}
\name{ttmoment-package}
\alias{ttmoment-package}
\title{ttmoment: Functional Trait Moment Calculation}
\description{
Calculates the community four ‘moments’(mean, variance, skewness, and kurtosis) of a given trait based on the moments described in Wieczynski et al. (2019) \doi{10.1073/pnas.1813723116}. These functional metrics are extremely useful in characterizing the distribution of traits in a plant community. Evaluating these community-weighted moments allows researchers to evaluate shifts in optimal phenotypes and understand the effect of selection and ecological filtering across various environments with exactness.

Calculates the community four ‘moments’(mean, variance, skewness, and kurtosis) of a given trait based on the moments described in Wieczynski et al. (2019) \doi{10.1073/pnas.1813723116}. These functional metrics are extremely useful in characterizing the distribution of traits in a plant community. Evaluating these community-weighted moments allows researchers to evaluate shifts in optimal phenotypes and understand the effect of selection and ecological filtering across various environments with exactness.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/ttmoment}
  \item \url{https://paulesantos.github.io/ttmoment/}
  \item Report bugs at \url{https://github.com/PaulESantos/ttmoment/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/ttmoment}
  \item \url{https://paulesantos.github.io/ttmoment/}
  \item Report bugs at \url{https://github.com/PaulESantos/ttmoment/issues}
}

}
\author{
\strong{Maintainer}: Paul Santos Andrade \email{paulefrens@gmail.com} (\href{https://orcid.org/0000-0002-6635-0375}{ORCID})

Other contributors:
\itemize{
  \item Macroecology Lab [copyright holder, funder]
}

}
\keyword{internal}
